forced re-render of navbar so that users don't see logged in nav bar post logout
